Move range to common code folder
